web_module/requirements.txt
Django
djangorestframework
pygments


